[DOCSET] rename stemma editor
